pam: do not use child pid after successful waitpid
If a (zombie) child process has been waited for, the PID can be
reused by the operating system for other processes. Set child to -1
to make sure that no received signal could be forwarded to a process
which received the same PID after our successful waitpid call.

pam: set SIGCHLD handling to default

If SIGCHLD is ignored by a user who calls doas, then no zombie process
is created when child process dies, which in turn means that calling
kill with value of "child" could send signals to wrong process.